+++ include/solarsystem.h
@@ -116,14 +116,14 @@ typedef short (*novas_planet_provider_hp)(const double jd_tdb[2], enum novas_pla
  * to identify the object for which ephemeris data is requested. As such you only need
  * to specify the one that is going to be used.
  *
+ * @param name          The name of the solar-system body (in case the ephemeris provider is
+ *                      name based, otherwiase it may be NULL).
  * @param id            The ID number of the solar-system body for which the position in
  *                      desired. (Typically a NAIF ID, or else an appropriate ID for the
  *                      implementation -- corresponding minor planet objects should be created
  *                      with the same type of ID.). If the ephemeris provider is name based
  *                      the ID is not used an can be set to anything (-1 might be a good
  *                      default).
- * @param name          The name of the solar-system body (in case the ephemeris provider is
- *                      name based, otherwiase it may be NULL).
  * @param jd_tdb_high   [day] The high-order part of Barycentric Dynamical Time (TDB) based
  *                      Julian date for which to find the position and velocity. Typically
  *                      this may be the integer part of the Julian date for high-precision
@@ -298,6 +298,13 @@ short planet_jplint(double jd_tdb, enum novas_planet body, enum novas_origin ori
 
 short planet_jplint_hp(const double jd_tdb[2], enum novas_planet body, enum novas_origin origin, double *position, double *velocity);
 
+
+// Added in v1.2 --------------------------------->
+
+novas_planet_provider get_planet_provider();
+
+novas_planet_provider_hp get_planet_provider_hp();
+
 #if USE_CALCEPH
 #  include "calceph.h"
 
@@ -308,6 +315,9 @@ int novas_use_calceph_planets(t_calcephbin *eph);
 #endif /* USE_CALCEPH */
 
 /// \cond PRIVATE
+
+
+// <================= SuperNOVAS internals ======================>
 
 #  ifdef __NOVAS_INTERNAL_API__